added desktop and mobile title for mission section
