+++ messages/uk.json
@@ -6,9 +6,9 @@
     "Mission": {
         "pageTitle": "наша місія",
         "titleFirstWord": "Створюємо ",
-        "titleSecondWord": "вражаючі",
+        "titleSecondWord": "вражаючі ",
         "titleThirdWord": "цифрові",
-        "titleFourthWord": " продукти",
+        "titleFourthWord": " продукти ",
         "titleFifthWord": "разом",
         "titleOtherWords": " з ідеальною командою",
         "firstBenefit": "Дизайн, розробка і тестування - це лише частина нашої роботи. Наші команди здатні на значно більше, коли разом.",
