+++ src/components/mission/Title/MobileTitle.tsx
@@ -1,0 +1,15 @@
+import { useTranslations } from "next-intl";
+
+export const MobileTitle = () => {
+    const getTranslation = useTranslations("Mission");
+    return (
+        <p className="pc:hidden">
+            {getTranslation("titleFirstWord")}
+            {getTranslation("titleSecondWord")}
+            {getTranslation("titleThirdWord")}
+            {getTranslation("titleFourthWord")}
+            {getTranslation("titleFifthWord")}
+            {getTranslation("titleOtherWords")}
+        </p>
+    );
+};
